Added AdvancedSettings window. Intermediate commit.
