+++ py_scripts/dbComparator/resources/sqlComparatorBacklog.txt
@@ -1,23 +1,23 @@
 Enchancements:
 
 1. Add try/except construction in case of defining non-existent columns/tables.
-2. We shouldn't run describe query on skipped tables.
-3. Comparing step should be moved from UI. Instead of manually possibility of setting comparing step I should implement
+2. Comparing step should be moved from UI. Instead of manually possibility of setting comparing step I should implement
 feature which automatically changes comparing step in case of slow or fast connection.
-4. Implement correct working of table timeout (probably, we should stop query, if timeout falls).
-5. Great UI refactoring - move different element groups to different classes.
-6. Add verification case when entities and reports flag disabled at same time.
-7. Disable checking mode group if "Only reports flag" disabled.
-8. UI hangs and disables during comparing DBs. Add modal window with progress bar and some information about
+3. Implement correct working of table timeout (probably, we should stop query, if timeout falls).
+4. Great UI refactoring - move different element groups to different classes.
+5. Add verification case when entities and reports flag disabled at same time.
+6. Disable checking mode group if "Only reports flag" disabled.
+7. UI hangs and disables during comparing DBs. Add modal window with progress bar and some information about
 progress of comparing. Probably, we should generate all sql-queries before running and use this queries to calculate
 comparing progress.
-9. Hide advanced settings. Probably, move it to different window.
-10. Probably remove Schema columns field, and move this data to hardcode.
+-> 8. Hide advanced settings. Probably, move it to different window.
+-> 9. Probably remove Schema columns field, and move this data to hardcode.
 
 Bugs:
 
 1. There are some fields in header of dumped file but only one field in data for page table on marvin.
 
 Epics:
+0. Support dataframe comparing.
 1. Present summary information more flexible.
 2. Probably, add functional for cloning databases.
